+++ src/lodge/lodge.service.ts
@@ -41,8 +41,11 @@ export class LodgeService {
     dto: HeaderLodgeUpdateDto,
     file?: Express.Multer.File,
   ) {
-    await this.filesService.delete(urlId);
-    const newImg = await this.filesService.upload(file, 'header');
+    let newImg: string;
+    if (urlId && file) {
+      await this.filesService.delete(urlId);
+      newImg = await this.filesService.upload(file, 'header');
+    }
     const [updatedCount, [updatedHeader]] = await this.headerModel.update(
       { ...dto, image: newImg },
       {
